+++ fractal_server/app/routes/api/v2/task_group.py
@@ -3,7 +3,6 @@ import itertools
 from fastapi import APIRouter
 from fastapi import Depends
 from fastapi import HTTPException
-from fastapi import Response
 from fastapi import status
 from packaging.version import InvalidVersion
 from packaging.version import parse
@@ -12,7 +11,6 @@ from pydantic.types import AwareDatetime
 from sqlmodel import or_
 from sqlmodel import select
 
-from ._aux_functions_task_lifecycle import check_no_ongoing_activity
 from ._aux_functions_tasks import _get_task_group_full_access
 from ._aux_functions_tasks import _get_task_group_read_access
 from ._aux_functions_tasks import _verify_non_duplication_group_constraint
@@ -23,7 +21,6 @@ from fractal_server.app.models import LinkUserGroup
 from fractal_server.app.models import UserOAuth
 from fractal_server.app.models.v2 import TaskGroupActivityV2
 from fractal_server.app.models.v2 import TaskGroupV2
-from fractal_server.app.models.v2 import WorkflowTaskV2
 from fractal_server.app.routes.auth import current_active_user
 from fractal_server.app.routes.auth._aux_auth import _get_default_usergroup_id
 from fractal_server.app.routes.auth._aux_auth import (
@@ -197,41 +194,6 @@ async def get_task_group(
         db=db,
     )
     return task_group
-
-
-@router.delete("/{task_group_id}/", status_code=204)
-async def delete_task_group(
-    task_group_id: int,
-    user: UserOAuth = Depends(current_active_user),
-    db: AsyncSession = Depends(get_async_db),
-):
-    """
-    Delete single TaskGroup
-    """
-
-    task_group = await _get_task_group_full_access(
-        task_group_id=task_group_id,
-        user_id=user.id,
-        db=db,
-    )
-
-    await check_no_ongoing_activity(task_group_id=task_group_id, db=db)
-
-    stm = select(WorkflowTaskV2).where(
-        WorkflowTaskV2.task_id.in_({task.id for task in task_group.task_list})
-    )
-    res = await db.execute(stm)
-    workflow_tasks = res.scalars().all()
-    if workflow_tasks != []:
-        raise HTTPException(
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            detail=f"TaskV2 {workflow_tasks[0].task_id} is still in use",
-        )
-
-    await db.delete(task_group)
-    await db.commit()
-
-    return Response(status_code=status.HTTP_204_NO_CONTENT)
 
 
 @router.patch("/{task_group_id}/", response_model=TaskGroupReadV2)