+++ fractal_server/tasks/v2/local/__init__.py
@@ -2,5 +2,6 @@ from .collect import collect_local  # noqa
 from .collect_pixi import collect_local_pixi  # noqa
 from .deactivate import deactivate_local  # noqa
 from .deactivate_pixi import deactivate_local_pixi  # noqa
+from .delete import delete_local  # noqa
 from .reactivate import reactivate_local  # noqa
 from .reactivate_pixi import reactivate_local_pixi  # noqa